requirements.txt
# list obtained from pip freeze, probably don't need all 
autopep8==1.2.4
cffi==1.7.0
cookies==2.2.1
cryptography==1.4
Django==1.9.9
django-allauth==0.23.0
django-bootstrap-form==3.2
django-bootstrap3==6.2.2
django-crispy-forms==1.4.0
django-filter==0.13.0
django-guardian==1.4.4
djangorestframework==3.4.1
enum34==1.1.6
flake8==2.5.4
funcsigs==1.0.2
geopy==1.11.0
idna==2.1
ipaddress==1.0.16
isort==4.2.5
jdcal==1.2
Markdown==2.6.6
mccabe==0.4.0
mock==2.0.0
ndg-httpsclient==0.4.2
numpy==1.11.1
oauthlib==1.1.2
pbr==1.10.0
pep8==1.7.0
pyasn1==0.1.9
pycparser==2.14
pyflakes==1.0.0
pyOpenSSL==16.0.0
pyowm==2.3.2
python-forecastio==1.3.4
python-openid==2.2.5
requests==2.7.0
requests-oauthlib==0.6.2
responses==0.5.1
six==1.10.0
